More `ByteBuffer` converters
