+++ buffer-netty/src/main/java/io/micronaut/buffer/netty/NettyByteBufferConverters.java
@@ -1,0 +1,89 @@
+/*
+ * Copyright 2017-2022 original authors
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package io.micronaut.buffer.netty;
+
+import io.micronaut.core.annotation.Internal;
+import io.micronaut.core.convert.MutableConversionService;
+import io.micronaut.core.convert.TypeConverterRegistrar;
+import io.micronaut.core.io.buffer.ByteBuffer;
+import io.netty.buffer.ByteBuf;
+import io.netty.buffer.ByteBufUtil;
+import io.netty.buffer.CompositeByteBuf;
+import io.netty.buffer.Unpooled;
+
+import java.util.Optional;
+
+/**
+ * Factory for {@link ByteBuf} and {@link NettyByteBuffer} related converters.
+ *
+ * @author Denis Stepanov
+ * @since 4.6
+ */
+@Internal
+public final class NettyByteBufferConverters implements TypeConverterRegistrar {
+    @Override
+    public void register(MutableConversionService conversionService) {
+
+        conversionService.addConverter(
+            NettyByteBuffer.class,
+            ByteBuf.class,
+            NettyByteBuffer::asNativeBuffer
+        );
+
+        conversionService.addConverter(
+            ByteBuf.class,
+            ByteBuffer.class,
+            NettyByteBuffer::new
+        );
+
+        conversionService.addConverter(
+            java.nio.ByteBuffer.class,
+            ByteBuf.class,
+            nioBuffer -> Unpooled.copiedBuffer(nioBuffer)
+        );
+
+        conversionService.addConverter(
+            ByteBuf.class,
+            CharSequence.class,
+            (byteBuf, target, context) -> Optional.of(byteBuf.toString(context.getCharset()))
+        );
+
+        conversionService.addConverter(
+            ByteBuf.class,
+            String.class,
+            (byteBuf, target, context) -> Optional.of(byteBuf.toString(context.getCharset()))
+        );
+
+        conversionService.addConverter(
+            CompositeByteBuf.class,
+            CharSequence.class,
+            (object, targetType, context) -> Optional.of(object.toString(context.getCharset()))
+        );
+
+        conversionService.addConverter(
+            ByteBuf.class,
+            byte[].class,
+            byteBuf -> ByteBufUtil.getBytes(byteBuf)
+        );
+
+        conversionService.addConverter(
+            byte[].class,
+            ByteBuf.class,
+            array -> Unpooled.wrappedBuffer(array)
+        );
+    }
+
+}
