+++ buffer-netty/src/main/resources/META-INF/services/io.micronaut.core.convert.TypeConverterRegistrar
@@ -1,0 +1,1 @@
+io.micronaut.buffer.netty.NettyByteBufferConverters
